Reverting back to straight quotes.
Curly quotes initially became question marks because the original encoding
of the script that created the verses was UTF-8 instead of Latin1.
However, when Latin1 was used, they became unreadable because BibleForge
was sending data to the client in UTF-8.  This could be solved by sending
a header entry to set the data to Latin1.  E.g.,

header('Content-type: application/json; charset=ISO-8859-1');

However, this would mean that other content (such as the original Greek
and Hebrew) would not be displayed properly.  Also, using an HTML entity
(i.e., &#x2019;) did not work either because the JavaScript highlighter
routine does not recognize that as punctuation and, therefore, does not
highlight those words.

A possible solution is to encode the English in Unicode.  This would work
but would probably noticeably slower.  Another solution would be to indicate
the encoding of each table and dynamically send the correct encoding in the
header.  The downside is that it is a bit more complicated and transfers
a little bit more data in the header.  This is probably the best solution
but will take some time to implement.